move env into it's own file
